// src/scripts/Model/Style.ts
export enum Style {
    NONE = 'NONE',
    MULTI_TASK_ONE_REQUIREMENT = 'MULTI_TASK_ONE_REQUIREMENT',
    MULTI_TASK_ONE_REQUIREMENT_WITH_IDEA = 'MULTI_TASK_ONE_REQUIREMENT_WITH_IDEA',
    ONE_TASK_MULTI_REQUIREMENTS = 'ONE_TASK_MULTI_REQUIREMENTS',
    ONE_TASK_MULTI_REQUIREMENTS_WITH_IDEA = 'ONE_TASK_MULTI_REQUIREMENTS_WITH_IDEA',
    ONE_TASK_ONE_REQUIREMENT = 'ONE_TASK_ONE_REQUIREMENT',
    ONE_TASK_ONE_REQUIREMENTS_WITH_IDEA = 'ONE_TASK_ONE_REQUIREMENTS_WITH_IDEA',
    IDEA_SELECTION = 'IDEA_SELECTION',

}